pyproject.toml
[tool.poetry]
name = "winoreat"
version = "0.1.0"
description = ""
authors = ["leegyurak <dev@devgyurak.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.3"
django = "^5.1.1"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
django-extensions = "^3.2.3"
django-filter = "^24.3"
django-cors-headers = "^4.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
